+++ lib/mgmg.rb
@@ -7,79 +7,106 @@ require_relative './mgmg/ir'
 require_relative './mgmg/system_equip'
 require_relative './mgmg/cuisine'
 require_relative './mgmg/reinforce'
+require_relative './mgmg/option'
 require_relative './mgmg/search'
 require_relative './mgmg/optimize'
 
 class String
-	def min_level(w=1)
-		Mgmg::Equip.min_level(self, w)
+	def min_weight(opt: Mgmg::Option.new)
+		build(build(opt: opt).min_levels_max, opt: opt).weight
 	end
-	def min_levels(w=1, left_associative: true)
-		build(-1, -1, left_associative: left_associative).min_levels(w)
+	def max_weight(include_outsourcing=false, opt: Mgmg::Option.new)
+		if include_outsourcing
+			build(-1, opt: opt).weight
+		else
+			build(min_smith(opt: opt), opt: opt).weight
+		end
 	end
-	def min_smith(left_associative: true)
-		Mgmg::Equip.min_smith(self, left_associative: left_associative)
+	def min_level(w=0, include_outsourcing=false, opt: Mgmg::Option.new)
+		built = build(-1, opt: opt)
+		w = build(built.min_levels_max, -1, opt: opt).weight - w if w <= 0
+		return -1 if include_outsourcing && built.weight <= w
+		ms = min_smith(opt: opt)
+		return ms if build(ms, opt: opt).weight <= w
+		ary = [ms]
+		4.downto(1) do |wi| # 単品の最大重量は[斧|重鎧](金10石10)の5
+			built.min_levels(wi).values.each do |v|
+				(ary.include?(v) or ary << v) if ms < v
+			end
+		end
+		ary.sort.each do |l|
+			return l if build(l, opt: opt).weight <= w
+		end
+		raise ArgumentError, "w=`#{w}' is given, but the minimum weight for the recipe is `#{min_weight(opt: opt)}'."
 	end
-	def min_comp(left_associative: true)
-		Mgmg::Equip.min_comp(self, left_associative: left_associative)
+	def min_levels(w=1, opt: Mgmg::Option.new)
+		build(opt: opt).min_levels(w)
 	end
-	def build(smith=-1, comp=smith, left_associative: true)
-		Mgmg::Equip.build(self, smith, comp, left_associative: left_associative)
+	def min_levels_max(w=1, opt: Mgmg::Option.new)
+		min_levels(w, opt: opt).values.append(-1).max
 	end
-	def ir(left_associative: true, reinforcement: [])
-		Mgmg::IR.build(self, left_associative: left_associative, reinforcement: reinforcement)
+	def min_smith(opt: Mgmg::Option.new)
+		Mgmg::Equip.min_smith(self, opt: opt)
 	end
-	def poly(para=:cost, left_associative: true)
-		la = left_associative
+	def min_comp(opt: Mgmg::Option.new)
+		Mgmg::Equip.min_comp(self, opt: opt)
+	end
+	def build(smith=-1, comp=smith, opt: Mgmg::Option.new)
+		Mgmg::Equip.build(self, smith, comp, left_associative: opt.left_associative).reinforce(*opt.reinforcement)
+	end
+	def ir(opt: Mgmg::Option.new)
+		Mgmg::IR.build(self, left_associative: opt.left_associative, reinforcement: opt.reinforcement)
+	end
+	def poly(para=:cost, opt: Mgmg::Option.new)
 		case para
 		when :atkstr
-			self.poly(:attack, left_associative: la) + self.poly(:str, left_associative: la)
+			self.poly(:attack, opt: opt) + self.poly(:str, opt: opt)
 		when :atk_sd
-			self.poly(:attack, left_associative: la) + self.poly(:str, left_associative: la).quo(2) + self.poly(:dex, left_associative: la).quo(2)
+			self.poly(:attack, opt: opt) + self.poly(:str, opt: opt).quo(2) + self.poly(:dex, opt: opt).quo(2)
 		when :dex_as
-			self.poly(:dex, left_associative: la) + self.poly(:attack, left_associative: la).quo(2) + self.poly(:str, left_associative: la).quo(2)
+			self.poly(:dex, opt: opt) + self.poly(:attack, opt: opt).quo(2) + self.poly(:str, opt: opt).quo(2)
 		when :mag_das
-			self.poly(:magic, left_associative: la) + self.poly(:dex_as, left_associative: la).quo(2)
+			self.poly(:magic, opt: opt) + self.poly(:dex_as, opt: opt).quo(2)
 		when :magmag
-			self.poly(:magdef, left_associative: la) + self.poly(:magic, left_associative: la).quo(2)
+			self.poly(:magdef, opt: opt) + self.poly(:magic, opt: opt).quo(2)
 		when :pmdef
-			pd = self.poly(:phydef, left_associative: la)
-			md = self.poly(:magmag, left_associative: la)
+			pd = self.poly(:phydef, opt: opt)
+			md = self.poly(:magmag, opt: opt)
 			pd <= md ? pd : md
 		when :cost
 			if Mgmg::SystemEquip.keys.include?(self)
 				return Mgmg::TPolynomial.new(Mgmg::Mat.new(1, 1, 0.quo(1)), 28, 0, 12, 12)
 			end
-			built = self.build(-1)
+			built = self.build(-1, opt: opt)
 			const = (built.star**2) * ( /\+/.match(self) ? 5 : ( built.kind < 8 ? 2 : 1 ) )
-			ret = poly(:attack, left_associative: la) + poly(:phydef, left_associative: la) + poly(:magdef, left_associative: la)
-			ret += poly(:hp, left_associative: la).quo(4) + poly(:mp, left_associative: la).quo(4)
-			ret += poly(:str, left_associative: la) + poly(:dex, left_associative: la) + poly(:speed, left_associative: la) + poly(:magic, left_associative: la)
+			ret = poly(:attack, opt: opt) + poly(:phydef, opt: opt) + poly(:magdef, opt: opt)
+			ret += poly(:hp, opt: opt).quo(4) + poly(:mp, opt: opt).quo(4)
+			ret += poly(:str, opt: opt) + poly(:dex, opt: opt) + poly(:speed, opt: opt) + poly(:magic, opt: opt)
 			ret.mat.body[0][0] += const
 			ret
 		else
-			Mgmg::TPolynomial.build(self, para, left_associative: la)
+			Mgmg::TPolynomial.build(self, para, left_associative: opt.left_associative)
 		end
 	end
-	def eff(para, smith, comp=smith, left_associative: true)
-		a = build(smith, comp, left_associative: left_associative).para_call(para)
-		b = build(smith+1, comp, left_associative: left_associative).para_call(para)
-		c = build(smith, comp+2, left_associative: left_associative).para_call(para)
+	def eff(para, smith, comp=smith, opt: Mgmg::Option.new)
+		a = build(smith, comp, opt: opt).para_call(para)
+		b = build(smith+1, comp, opt: opt).para_call(para)
+		c = build(smith, comp+2, opt: opt).para_call(para)
 		sden = smith==0 ? 1 : 2*smith-1
 		cden = comp==0 ? 4 : 8*comp
 		[(b-a).quo(sden), (c-a).quo(cden)]
 	end
-	def peff(para, smith, comp=smith, left_associative: true)
-		poly(para, left_associative: left_associative).eff(smith, comp)
+	def peff(para, smith, comp=smith, opt: Mgmg::Option.new)
+		poly(para, opt: opt).eff(smith, comp)
 	end
-	def show(smith=-1, comp=smith, left_associative: true, para: :power, reinforcement: [])
-		rein = case reinforcement
+	def show(smith=-1, comp=smith, para: :power, opt: Mgmg::Option.new)
+		rein = case opt.reinforcement
 		when Array
-			reinforcement.map{|r| Mgmg::Reinforcement.compile(r)}
+			opt.reinforcement.map{|r| Mgmg::Reinforcement.compile(r)}
 		else
-			[Mgmg::Reinforcement.compile(reinforcement)]
+			[Mgmg::Reinforcement.compile(opt.reinforcement)]
 		end
-		built = self.build(smith, comp, left_associative: left_associative).reinforce(*rein)
+		built = build(smith, comp, opt: opt)
 		pstr = '%.3f' % built.para_call(para)
 		pstr.sub!(/\.?0+\Z/, '')
 		puts "Building"
@@ -88,37 +115,39 @@ class String
 		puts "with levels (#{smith}, #{comp})#{rein} yields (#{pstr}, #{built.total_cost})"
 		puts "  #{built}"
 	end
-	def phydef_optimize(smith=nil, comp=smith, left_associative: true, magdef_maximize: true)
-		Mgmg::Optimize.phydef_optimize(self, smith, comp, left_associative: left_associative, magdef_maximize: magdef_maximize)
+	def phydef_optimize(smith=nil, comp=smith, opt: Mgmg::Option.new)
+		Mgmg::Optimize.phydef_optimize(self, smith, comp, opt: opt)
 	end
-	def buster_optimize(smith=nil, comp=smith, left_associative: true)
-		Mgmg::Optimize.buster_optimize(self, smith, comp, left_associative: left_associative)
+	def buster_optimize(smith=nil, comp=smith, opt: Mgmg::Option.new)
+		Mgmg::Optimize.buster_optimize(self, smith, comp, opt: opt)
 	end
 end
 module Enumerable
-	def build(smith=-1, armor=smith, comp=armor.tap{armor=smith}, left_associative: true)
-		self.sum do |str|
-			m = /\A\[*([^\+]+)/.match(str)
-			if Mgmg::EquipPosition[m[1].build(0).kind] == 0
-				str.build(smith, comp, left_associative: left_associative)
+	def build(smith=-1, armor=smith, comp=armor.tap{armor=smith}, opt: Mgmg::Option.new)
+		opt = opt.dup
+		rein = opt.reinforcement
+		opt.reinforcement = []
+		self.sum(Mgmg::Equip::Zero) do |str|
+			if Mgmg::EquipPosition[str.build(opt: opt).kind] == 0
+				str.build(smith, comp, opt: opt)
 			else
-				str.build(armor, comp, left_associative: left_associative)
+				str.build(armor, comp, opt: opt)
 			end
-		end
+		end.reinforce(*rein)
 	end
-	def ir(left_associative: true, reinforcement: [])
-		self.sum do |str|
-			str.ir(left_associative: left_associative)
-		end.add_reinforcement(reinforcement)
+	def ir(opt: Mgmg::Option.new)
+		self.sum(Mgmg::IR::Zero) do |str|
+			str.ir(opt: opt)
+		end.add_reinforcement(opt.reinforcement)
 	end
-	def show(smith=-1, armor=smith, comp=armor.tap{armor=smith}, left_associative: true, para: :power, reinforcement: [])
-		rein = case reinforcement
+	def show(smith=-1, armor=smith, comp=armor.tap{armor=smith}, para: :power, opt: Mgmg::Option.new)
+		rein = case opt.reinforcement
 		when Array
-			reinforcement.map{|r| Mgmg::Reinforcement.compile(r)}
+			opt.reinforcement.map{|r| Mgmg::Reinforcement.compile(r)}
 		else
-			[Mgmg::Reinforcement.compile(reinforcement)]
+			[Mgmg::Reinforcement.compile(opt.reinforcement)]
 		end
-		built = self.build(smith, armor, comp, left_associative: left_associative).reinforce(*rein)
+		built = self.build(smith, armor, comp, opt: opt)
 		pstr = '%.3f' % built.para_call(para)
 		pstr.sub!(/\.?0+\Z/, '')
 		puts "Building"
@@ -127,14 +156,75 @@ module Enumerable
 		puts "with levels (#{smith}, #{armor}, #{comp})#{rein} yields (#{pstr}, #{built.total_cost})"
 		puts "  #{built}"
 	end
-	def min_levels(w=1, left_associative: true)
-		build(-1, -1, -1, left_associative: left_associative).min_levels(w)
+	def min_weight(opt: Mgmg::Option.new)
+		build(*build(opt: opt).min_levels_max, -1, opt: opt).weight
 	end
-	def min_level(w=1, left_associative: true)
-		ret = [0, 0]
-		build(-1, -1, -1, left_associative: left_associative).min_levels(w).each do |str, level|
-			m = /\A\[*([^\+]+)/.match(str)
-			if Mgmg::EquipPosition[m[1].build(0).kind] == 0
+	def max_weight(include_outsourcing=false, opt: Mgmg::Option.new)
+		if include_outsourcing
+			build(-1, opt: opt).weight
+		else
+			build(*min_smith(opt: opt), -1, opt: opt).weight
+		end
+	end
+	def min_weights(opt: Mgmg::Option.new)
+		weapons, armors = [], []
+		each do |str|
+			if Mgmg::EquipPosition[str.build(opt: opt).kind] == 0
+				weapons << str
+			else
+				armors << str
+			end
+		end
+		[weapons.min_weight(opt: opt), armors.min_weight(opt: opt)]
+	end
+	def max_weights(include_outsourcing=false, opt: Mgmg::Option.new)
+		weapons, armors = [], []
+		each do |str|
+			if Mgmg::EquipPosition[str.build(opt: opt).kind] == 0
+				weapons << str
+			else
+				armors << str
+			end
+		end
+		[weapons.max_weight(include_outsourcing, opt: opt), armors.max_weight(include_outsourcing, opt: opt)]
+	end
+	def min_level(ws=0, wa=ws, include_outsourcing=false, opt: Mgmg::Option.new)
+		weapons, armors = [], []
+		each do |str|
+			if Mgmg::EquipPosition[str.build(opt: opt).kind] == 0
+				weapons << str
+			else
+				armors << str
+			end
+		end
+		ms, ma = min_smith(opt: opt)
+		rs = min_level_sub(ws, ms, 0, weapons, include_outsourcing, opt: opt)
+		ra = min_level_sub(wa, ma, 1, armors, include_outsourcing, opt: opt)
+		[rs, ra]
+	end
+	private def min_level_sub(w, ms, i, recipe, include_outsourcing, opt: Mgmg::Option.new)
+		built = recipe.build(opt: opt)
+		w = recipe.build(built.min_levels_max[i], opt: opt).weight - w if w <= 0
+		return -1 if include_outsourcing && built.weight <= w
+		return ms if build(ms, opt: opt).weight <= w
+		ary = [ms]
+		4.downto(1) do |wi|
+			built.min_levels(wi).values.each do |v|
+				(ary.include?(v) or ary << v) if ms << v
+			end
+		end
+		ary.sort.each do |l|
+			return l if recipe.build(l, opt: opt).weight <= w
+		end
+		raise ArgumentError, "w#{%w|s a|[i]}=`#{w}' is given, but the minimum weight for the #{%w|weapon(s) armor(s)|[i]} is `#{recipe.min_weight(opt: opt)}'."
+	end
+	def min_levels(w=1, opt: Mgmg::Option.new)
+		build(opt: opt).min_levels(w)
+	end
+	def min_levels_max(w=1, opt: Mgmg::Option.new)
+		ret = [-1, -1]
+		min_levels(w, opt: opt).each do |str, level|
+			if Mgmg::EquipPosition[str.build(opt: opt).kind] == 0
 				ret[0] = [ret[0], level].max
 			else
 				ret[1] = [ret[1], level].max
@@ -142,12 +232,11 @@ module Enumerable
 		end
 		ret
 	end
-	def min_smith(left_associative: true)
-		ret = [0, 0]
+	def min_smith(opt: Mgmg::Option.new)
+		ret = [-1, -1]
 		self.each do |str|
-			s = Mgmg::Equip.min_smith(str, left_associative: left_associative)
-			m = /\A\[*([^\+]+)/.match(str)
-			if Mgmg::EquipPosition[m[1].build(0).kind] == 0
+			s = Mgmg::Equip.min_smith(str, opt: opt)
+			if Mgmg::EquipPosition[str.build(opt: opt).kind] == 0
 				ret[0] = [ret[0], s].max
 			else
 				ret[1] = [ret[1], s].max
@@ -155,9 +244,9 @@ module Enumerable
 		end
 		ret
 	end
-	def min_comp(left_associative: true)
+	def min_comp(opt: Mgmg::Option.new)
 		self.map do |str|
-			Mgmg::Equip.min_comp(str, left_associative: left_associative)
-		end.max
+			Mgmg::Equip.min_comp(str, opt: opt)
+		end.append(-1).max
 	end
 end